+++ src/main/java/fr/inrae/act/bagap/chloe/kernel/AreaCountCoupleKernel.java
@@ -6,10 +6,10 @@ public class AreaCountCoupleKernel extends AreaLandscapeMetricKernel{
 	
 	private final int[] mapValues;
 
-	public AreaCountCoupleKernel(int noDataValue, short[] values){
+	public AreaCountCoupleKernel(int noDataValue, int[] values){
 		super(noDataValue);
 		int maxV = 0;
-		for(short v : values){
+		for(int v : values){
 			maxV = Math.max(v, maxV);
 		}
 		maxV++;
@@ -19,13 +19,13 @@ public class AreaCountCoupleKernel extends AreaLandscapeMetricKernel{
 		}
 		mapCouples = new int[values.length][values.length];
 		int index = 0;
-		for(short v : values){
+		for(int v : values){
 			mapCouples[mapValues[v]][mapValues[v]] = index;
 			index++;
 		}
 		
-		for(short v1 : values){
-			for(short v2 : values){
+		for(int v1 : values){
+			for(int v2 : values){
 				if(v1 < v2) {
 					mapCouples[mapValues[v1]][mapValues[v2]] = index;
 					mapCouples[mapValues[v2]][mapValues[v1]] = index;
